requirements.txt
python-dotenv
requests
fastapi
uvicorn
openai==1.52.0

